disasm (basic)
Disassemble with elfutils. The elfutils library have not yet implemented
disassemblers other than i386/x86-64, but the interface is
platform-independent and it doesn't introduce another dependency as
backend is already using libelf.

Elfutils loads some additional libraries from libebl, which relies on
the location of executable. It normally won't work if executable is not
located in /usr/bin, and the workaround is to use RUNPATH in libbackend.
Also SConscripts are running CC to check library and header path, so we
are not afraid if elfutils is installed in non-standard locations.

BFD is another choice for disassembler, and it can also parse ELF
itself. Maybe write another backend with BFD?